+++ cobib/commands/open.py
@@ -19,15 +19,15 @@ class OpenCommand(Command):
     name = 'open'
 
     def execute(self, args, out=sys.stderr):
-        """Open file from entry.
+        """Open file from entries.
 
-        Opens the associated file of an entry with xdg-open.
+        Opens the associated file of the entries with xdg-open.
 
         Args: See base class.
         """
         LOGGER.debug('Starting Open command.')
         parser = ArgumentParser(prog="open", description="Open subcommand parser.")
-        parser.add_argument("label", type=str, help="label of the entry")
+        parser.add_argument("labels", type=str, nargs='+', help="labels of the entries")
 
         if not args:
             parser.print_usage(sys.stderr)
@@ -39,47 +39,51 @@ class OpenCommand(Command):
             print("{}: {}".format(exc.argument_name, exc.message), file=sys.stderr)
             return None
 
-        try:
-            entry = CONFIG.config['BIB_DATA'][largs.label]
-            if 'file' not in entry.data.keys() or entry.data['file'] is None:
-                msg = "Error: There is no file associated with this entry."
-                LOGGER.error(msg)
-                if out is None:
-                    # called from TUI
-                    return msg
-                sys.exit(1)
-            opener = None
-            opener = CONFIG.config['DATABASE'].get('open')
+        errors = []
+        for label in largs.labels:
             try:
-                LOGGER.debug('Parsing "%s" for URLs.', entry.data['file'])
-                url = urlparse(entry.data['file'])
-                if url.scheme:
-                    # actual URL
-                    url = url.geturl()
-                else:
-                    # assume we are talking about a file and thus get its absolute path
-                    url = os.path.abspath(url.geturl())
-                LOGGER.debug('Opening "%s" with %s.', url, opener)
-                err = subprocess.Popen([opener, url], stderr=subprocess.PIPE)
-                err = err.stderr.read().decode()
-                if err:
-                    raise FileNotFoundError(err)
-            except FileNotFoundError as err:
-                LOGGER.error(err)
-                return str(err)
-        except KeyError:
-            msg = "Error: No entry with the label '{}' could be found.".format(largs.label)
-            LOGGER.error(msg)
+                entry = CONFIG.config['BIB_DATA'][label]
+                if 'file' not in entry.data.keys() or entry.data['file'] is None:
+                    msg = "Error: There is no file associated with this entry."
+                    LOGGER.error(msg)
+                    if out is None:
+                        errors.append(msg)
+                    continue
+                opener = CONFIG.config['DATABASE'].get('open', None)
+                try:
+                    LOGGER.debug('Parsing "%s" for URLs.', entry.data['file'])
+                    url = urlparse(entry.data['file'])
+                    if url.scheme:
+                        # actual URL
+                        url = url.geturl()
+                    else:
+                        # assume we are talking about a file and thus get its absolute path
+                        url = os.path.abspath(url.geturl())
+                    LOGGER.debug('Opening "%s" with %s.', url, opener)
+                    with open(os.devnull, 'w') as devnull:
+                        subprocess.Popen([opener, url], stdout=devnull, stderr=devnull,
+                                         stdin=devnull, close_fds=True)
+                except FileNotFoundError as err:
+                    LOGGER.error(err)
+                    errors.append(str(err))
+            except KeyError:
+                msg = "Error: No entry with the label '{}' could be found.".format(label)
+                LOGGER.error(msg)
 
-        return None
+        return '\n'.join(errors)
 
     @staticmethod
     def tui(tui):
         """See base class."""
         LOGGER.debug('Open command triggered from TUI.')
-        # get current label
-        label, _ = tui.get_current_label()
+        if tui.selection:
+            # use selection for command
+            labels = list(tui.selection)
+        else:
+            # get current label
+            label, _ = tui.get_current_label()
+            labels = [label]
         # populate buffer with entry data
-        error = OpenCommand().execute([label], out=None)
+        error = OpenCommand().execute(labels, out=None)
         if error:
             tui.prompt_print(error)